refactor: ProductRepository 추상화
